requirements.txt
#Flask==1.1.2
Flask==2.1.0
Flask-Caching==2.0.1
pandas==1.4.2
#jwt==1.3.1
PyJWT==2.6.0
#jinja2==3.1.2